notes delet works now. Also pulled notes out of indiv trip page for easier reading.
