+++ react-app/src/components/NoteCards/NoteCards.css
@@ -68,25 +68,86 @@ input {
     background-color: #ffffff;
 }
 
-
-
-
-
-.addTripButton,
+.addNoteButton,
 .deleteNoteButton {
     border-bottom: 4px solid transparent;
     display:inline;
-    width: 110px;
     margin: 0 auto;
 }
 
-.addTripButton.active,
-.addTripButton:hover {
+.addNoteButton.active,
+.addNoteButton:hover {
     border-color: #33CCFF;
 }
-
 
 .deleteNoteButton.active,
 .deleteNoteButton:hover {
     border-color: #f63d43;
+}
+
+
+.formContainer7 {
+    height: auto;
+    width: 400px;
+    background-color: rgba(255,255,255,0.13);
+    position: absolute;
+    transform: translate(-50%,-50%);
+    top: 50%;
+    left: 50%;
+    border-radius: 10px;
+    backdrop-filter: blur(10px);
+    border: 2px solid rgba(255,255,255,0.1);
+    box-shadow: 0 0 40px rgba(8,7,16,0.6);
+    padding: 50px 35px;
+}
+
+.formContainer7  *{
+    font-family: 'Poppins',sans-serif;
+    color: #ffffff;
+    letter-spacing: 0.5px;
+    outline: none;
+    border: none;
+}
+
+.formContainer7  h3{
+    font-size: 26px;
+    font-weight: 500;
+    line-height: 42px;
+    text-align: center;
+}
+
+label {
+    display: block;
+    margin-top: 14px;
+    font-size: 16px;
+    font-weight: 500;
+}
+
+input {
+    display: block;
+    height: 50px;
+    width: 100%;
+    background-color: rgba(255,255,255,0.07);
+    border-radius: 3px;
+    padding: 0 10px;
+    margin-top: 8px;
+    font-size: 14px;
+    font-weight: 300;
+}
+
+#delete {
+    margin-top: 50px;
+    width: 100%;
+    color: #ffffff;
+    padding: 15px 0;
+    font-size: 18px;
+    font-weight: 600;
+    border-radius: 5px;
+    cursor: pointer;
+}
+
+#delete:hover,
+#delete:active {
+    color: #080710;
+    background-color: #ffffff;
 }
