+++ react-app/src/store/invited_user.js
@@ -36,7 +36,6 @@ const deleteInvitedUsers = (id) => {
 
 export const postInvitedUsers = (tripAndUserName) => async (disptach) => {
     const {tripId, userName} = tripAndUserName
-    console.log("THIS IS WHETHER OR NOT THIS IS A KEY VALUE PAIR OR NOT-----------------", tripId)
     const response = await fetch('/api/invited_users/', { 
         method: 'POST',
         headers: { 'Content-Type': 'application/json' },
@@ -47,13 +46,6 @@ export const postInvitedUsers = (tripAndUserName) => async (disptach) => {
         const data = await response.json();
         // data = {invitedUser: {email: 'marnie@aa.io', id: 2, username: 'marnie'}, tripId: '1'}
         let invitedUserId = data.invitedUser.id
-
-        console.log("THIS IS DATA FOR USERNAME IN THE STORE-------------------", data)
-        console.log("ThIS IS INVITED USER ID FOR THE STORE------", invitedUserId)
-        
-        // let integerInvitedUserId = parseInt(invitedUserId, 10)
-        // // data gives us username
-        // // query here for trip id and user id 
         const response2 = await fetch(`/api/trips/${tripId}/users`, {
             method: "POST",
             headers: {'Content-Type': 'application/json'},
@@ -75,10 +67,8 @@ export const postInvitedUsers = (tripAndUserName) => async (disptach) => {
 
 export const loadInvitedUsers = (tripId) => async (dispatch) => {
     const res = await fetch(`/api/trips/${tripId}/users`)
-    // const res = await fetch(`/api/trips/users/${userId}/`) // thinking we dont need the trailing slashes
     if (res.ok) {
         const data = await res.json();
-        console.log("THIS IS INVITED USERS FOR THE FRONTEND FOR GET---------------", data)
         dispatch(getInvitedUsers(data))
     }
 }
